+++ formatter/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.1.5
+
+- Prevents email addresses from becoming directives (#21)
+
 ## 1.1.4
 
 - Prevents email addresses from becoming directives (#21)